Documented every class and fixed some minor bugs

+++ src/engine/Methods.java
@@ -1,7 +1,12 @@
 package engine;
 
+/*
+ * Clamps a value between min and max
+ * @author MCForsas 2018
+ */
+
 public class Methods {
-	public static int clamp(int value, int minValue, int maxValue) {
+	public static float clamp(float value, float minValue, float maxValue) {
 		if(value <= minValue) {
 			return minValue;
 		}else if(value >= maxValue) {
